feat: Added the spark_dedpup_dag
The DAG reads in consecutive .parquet files produced by the otodom_scraping_dag, and combines them into a single .parquet file
that contains only unique (in terms of the "link" column) observations.
This is done using the
add_increment
function that does an upsert into the final frame (merge + deduplication).

Resolves: #10